+++ tests/test_templatetags_extends.py
@@ -1075,3 +1075,70 @@ class TestExtendsCompat:
             <p data-djc-id-ca1bc40 data-djc-id-ca1bc42>This template extends another template.</p>
         """,
         )
+
+    @djc_test(parametrize=PARAMETRIZE_CONTEXT_BEHAVIOR)
+    def test_double_include_template_with_extend(
+        self,
+        components_settings,
+    ):
+        @register("simple_component")
+        class SimpleComponent(Component):
+            template: types.django_html = """
+                {% slot 'content' / %}
+            """
+
+        # Confirm that this setup works in Django without components
+        template1: types.django_html = """
+            {% extends 'block.html' %}
+            {% load component_tags %}
+            {% block body %}
+                {# history: [<Origin name='/Users/presenter/repos/django-components/tests/templates/included.html'>] #}
+                {% include 'included.html' with variable="INCLUDED 1" %}
+                {# history: [<Origin name='/Users/presenter/repos/django-components/tests/templates/included.html'>] #}
+                {% include 'included.html' with variable="INCLUDED 2" %}
+            {% endblock %}
+        """
+        rendered1 = Template(template1).render(Context())
+        expected1 = """
+            <!DOCTYPE html>
+            <html lang="en">
+              <body>
+                <main role="main">
+                  <div class='container main-container'>
+                    Variable: <strong>INCLUDED 1</strong>
+                    Variable: <strong>INCLUDED 2</strong>
+                  </div>
+                </main>
+              </body>
+            </html>
+        """
+        assertHTMLEqual(rendered1, expected1)
+
+        template2: types.django_html = """
+            {% extends 'block.html' %}
+            {% load component_tags %}
+            {% block body %}
+                {% component "simple_component" %}
+                    {% fill "content" %}
+                        {% include 'included.html' with variable="INCLUDED 1" %}
+                        {% include 'included.html' with variable="INCLUDED 2" %}
+                    {% endfill %}
+                {% endcomponent %}
+            {% endblock %}
+        """
+        rendered2 = Template(template2).render(Context({"DJC_DEPS_STRATEGY": "ignore"}))
+
+        expected2 = """
+            <!DOCTYPE html>
+            <html lang="en">
+              <body>
+                <main role="main">
+                  <div class='container main-container'>
+                    Variable: <strong data-djc-id-ca1bc40="">INCLUDED 1</strong>
+                    Variable: <strong data-djc-id-ca1bc40="">INCLUDED 2</strong>
+                  </div>
+                </main>
+              </body>
+            </html>
+        """
+        assertHTMLEqual(rendered2, expected2)